rename DefaultHeaderMap to HashBackedHeaderMap
while the current header map implementation remains the default,
and indeed only available imlpementation of HeaderMap, it's no longer
called DefaultHeaderMap, but HashBackedHeaderMap, so the default could
change more easily
